refactor: rename package name from data to dto
