+++ src/main/kotlin/com/kroffle/knitting/infra/calculator/PatternCalculator.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.kroffle.knitting.domain.design.value.Gauge
 import com.kroffle.knitting.domain.design.value.Pattern
-import com.kroffle.knitting.infra.calculator.data.toDraftPattern
+import com.kroffle.knitting.infra.calculator.dto.toDraftPattern
 
 class PatternCalculator {
     private val mapper: ObjectMapper = ObjectMapper().registerKotlinModule()
