--- src/main/kotlin/com/kroffle/knitting/infra/calculator/data/GaugeType.kt
@@ -1,6 +1,0 @@
-package com.kroffle.knitting.infra.calculator.data
-
-enum class GaugeType {
-    Stitch,
-    Row
-}
