+++ src/main/kotlin/com/kroffle/knitting/infra/calculator/dto/CalculateType.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.infra.calculator.data
+package com.kroffle.knitting.infra.calculator.dto
 
 enum class CalculateType(
     val gaugeType: GaugeType,
