+++ src/main/kotlin/com/kroffle/knitting/infra/calculator/dto/GaugeType.kt
@@ -1,0 +1,6 @@
+package com.kroffle.knitting.infra.calculator.dto
+
+enum class GaugeType {
+    Stitch,
+    Row
+}
